+++ include/IIntersectStrategy.h
@@ -8,10 +8,10 @@ class IIntersectStrategy {
 public:
 	// intersection with object
 	virtual void UpdateInter(Intersection& inter, Scene& sce,
-		const Vector3f& rayOrig, const Vector3f& rayDir) = 0;
+		const Vector3d& rayOrig, const Vector3d& rayDir) = 0;
 
 	// calculate shadow coefficient, for hard shadow
-	virtual float getShadowCoeffi(Scene& sce, Intersection& p, Vector3f& lightpos) = 0;
+	virtual double getShadowCoeffi(Scene& sce, Intersection& p, Vector3d& lightpos) = 0;
 
 
 };
