+++ include/Sphere.hpp
@@ -5,16 +5,16 @@
 
 class Sphere : public Object{
 public:
-	Vector3f centerPos;		// center position
-	float radius;
+	Vector3d centerPos;		// center position
+	double radius;
 
 	Sphere() {
-		centerPos = Vector3f(0.f, 0.f, 0.f);
-		radius = 1.f;
+		centerPos = Vector3d(0.0, 0.0, 0.0);
+		radius = 1.0;
 	}
 
-	Sphere(float x, float y, float z, float r) {
-		centerPos = Vector3f(x, y, z);
+	Sphere(double x, double y, double z, double r) {
+		centerPos = Vector3d(x, y, z);
 		radius = r;
 	}
 
@@ -23,15 +23,15 @@ public:
 
 	// check if the ray will intersect with this sphere or not
 	// if true, then set the nearest time of intersection as tNear, and mtlColor
-	bool intersect(const Vector3f& orig, const Vector3f& dir, Intersection& inter) override {
-		float A = 1.f;	// here is 1 since we are using normalized vector
-		float B = 2 * (dir.x * (orig.x - centerPos.x) + dir.y * (orig.y - centerPos.y)
+	bool intersect(const Vector3d& orig, const Vector3d& dir, Intersection& inter) override {
+		double A = 1.0;	// here is 1 since we are using normalized vector
+		double B = 2 * (dir.x * (orig.x - centerPos.x) + dir.y * (orig.y - centerPos.y)
 			+ dir.z * (orig.z - centerPos.z));
-		float C = pow((orig.x - centerPos.x), 2) + pow((orig.y - centerPos.y), 2)
+		double C = pow((orig.x - centerPos.x), 2) + pow((orig.y - centerPos.y), 2)
 			+ pow((orig.z - centerPos.z), 2) - radius * radius;
 
-		float t1 = 0;
-		float t2 = 0;
+		double t1 = 0;
+		double t2 = 0;
 		solveQuadratic(t1, t2, A, B, C);
 
 		inter.intersected = false;
@@ -56,19 +56,19 @@ public:
 				if (isTextureActivated)
 				{
 					// calculate the uv coordinate of this intersection
-					float u, v;
-					float phi = acos(inter.nDir.z);
+					double u, v;
+					double phi = acos(inter.nDir.z);
 					v = phi / M_PI;
 
-					float theta = atan2(inter.nDir.y, inter.nDir.x);	// return [-pi, pi]
+					double theta = atan2(inter.nDir.y, inter.nDir.x);	// return [-pi, pi]
 					// we need to map it to [0, 1]
 					if (theta < 0) theta += 2 * M_PI;	// trigonometric functions are periodic
-					u = (theta / (2.f * M_PI));			// 0 + [0, 1]    then if theta == 0, it is the left most point in width
+					u = (theta / (2.0 * M_PI));			// 0 + [0, 1]    then if theta == 0, it is the left most point in width
 
 					// or 
-					// u = 0.5 + (theta / (2.f * M_PI));  // 0.5 + [-0.5, 0.5]	  then if theta == 0, it is the middle point in width
+					// u = 0.5 + (theta / (2.0 * M_PI));  // 0.5 + [-0.5, 0.5]	  then if theta == 0, it is the middle point in width
 
-					inter.textPos = Vector2f(u, v);
+					inter.textPos = Vector2d(u, v);
 					inter.diffuseIndex = this->textureIndex;
 					inter.normalMapIndex = normalMapIndex;
 					inter.roughnessMapIndex = roughnessMapIndex;
@@ -100,19 +100,19 @@ public:
 			if (isTextureActivated)
 			{
 				// calculate the uv coordinate of this intersection
-				float u, v;
-				float phi = acos(inter.nDir.z);	// return [0, pi]
+				double u, v;
+				double phi = acos(inter.nDir.z);	// return [0, pi]
 				v = phi / M_PI;
 
-				float theta = atan2(inter.nDir.y, inter.nDir.x);	// return [-pi, pi]
+				double theta = atan2(inter.nDir.y, inter.nDir.x);	// return [-pi, pi]
 				// we need to map it to [0, 1]
 				if (theta < 0) theta += 2 * M_PI;	// trigonometric functions are periodic
-				u = (theta / (2.f * M_PI));			// 0 + [0, 1]    then if theta == 0, it is the left most point in width
+				u = (theta / (2.0 * M_PI));			// 0 + [0, 1]    then if theta == 0, it is the left most point in width
 
 				// or 
-				// u = 0.5 + (theta / (2.f * M_PI));  // 0.5 + [-0.5, 0.5]	  then if theta == 0, it is the middle point in width
+				// u = 0.5 + (theta / (2.0 * M_PI));  // 0.5 + [-0.5, 0.5]	  then if theta == 0, it is the middle point in width
 
-				inter.textPos = Vector2f(u, v);
+				inter.textPos = Vector2d(u, v);
 				inter.diffuseIndex = this->textureIndex;
 				inter.normalMapIndex = normalMapIndex;
 				inter.roughnessMapIndex = roughnessMapIndex;
@@ -125,16 +125,16 @@ public:
 
 
 	void initializeBound() override {
-		Vector3f min = { centerPos.x - radius, centerPos.y - radius, centerPos.z - radius };
-		Vector3f max = { centerPos.x + radius, centerPos.y + radius, centerPos.z + radius };
+		Vector3d min = { centerPos.x - radius, centerPos.y - radius, centerPos.z - radius };
+		Vector3d max = { centerPos.x + radius, centerPos.y + radius, centerPos.z + radius };
 		bound = BoundBox(min, max);
 	}
 
-	float getArea() override {
+	double getArea() override {
 		return radius * radius * M_PI;
 	}
 
-	void samplePoint(Intersection& inter, float& pdf) override {
+	void samplePoint(Intersection& inter, double& pdf) override {
 		// 10/1/2023:
 		// is it correct? maybe I should Inverse Trasnform sampling?
 
@@ -142,8 +142,8 @@ public:
 		// phi: angle from +z to xy plane
 		// 7/6/2023: 
 
-		float theta = getRandomFloat() * 2 * M_PI;
-		float phi = getRandomFloat() * M_PI;
+		double theta = getRandomFloat() * 2 * M_PI;
+		double phi = getRandomFloat() * M_PI;
 
 		inter.pos.x = centerPos.x + radius * cos(theta) * sin(phi);
 		inter.pos.y = centerPos.y + radius * sin(theta) * sin(phi);
